requirements.txt
ultralytics
fastapi
uvicorn
pydantic
opencv-python
easyocr
pillow
python-multipart